+++ conf/aws/application.conf
@@ -72,7 +72,7 @@ gobblin.aws.nfs.parent.dir="/home/ec2-user"
 gobblin.aws.nfs.root.dir=${gobblin.aws.nfs.parent.dir}"/"${gobblin.aws.cluster.name}
 gobblin.cluster.helix.cluster.name=${gobblin.aws.cluster.name}
 gobblin.cluster.job.conf.path=${gobblin.aws.nfs.root.dir}"/job-conf"
-gobblin.cluster.job.configuration.manager="gobblin.aws.AWSJobConfigurationManager"
+gobblin.cluster.job.configuration.manager= "gobblin2.aws.AWSJobConfigurationManager"
 gobblin.aws.email.notification.on.shutdown=false
 
 ## AWS Configuration
