+++ conf/log4j-compaction.xml
@@ -29,14 +29,14 @@
 
   <appender name="file" class="org.apache.log4j.RollingFileAppender">
     <param name="append" value="true" />
-    <param name="file" value="${gobblin.logs.dir}/gobblin-compaction.log" />
+    <param name="file" value="${gobblin2.logs.dir}/gobblin-compaction.log" />
     <layout class="org.apache.log4j.PatternLayout">
       <param name="ConversionPattern"
         value="%d{yyyy-MM-dd HH:mm:ss z} %-5p [%t] %C %X{tableName} %L - %m%n" />
     </layout>
   </appender>
 
-  <logger name="gobblin.compaction" additivity="false">
+  <logger name="gobblin2.compaction" additivity="false">
     <level value="info" />
     <appender-ref ref="console" />
   </logger>
