+++ conf/log4j-mapreduce.xml
@@ -21,12 +21,12 @@
 <log4j:configuration>
 
   <appender name="FileRoll" class="org.apache.log4j.rolling.RollingFileAppender">
-    <param name="file" value="${gobblin.logs.dir}/gobblin-current.log" />
+    <param name="file" value="${gobblin2.logs.dir}/gobblin-current.log" />
     <param name="append" value="true" />
     <param name="encoding" value="UTF-8" />
 
     <rollingPolicy class="org.apache.log4j.rolling.TimeBasedRollingPolicy">
-      <param name="FileNamePattern" value="${gobblin.logs.dir}/archive/gobblin.%d{yyyy-MM-dd}.log"/>
+      <param name="FileNamePattern" value="${gobblin2.logs.dir}/archive/gobblin2.%d{yyyy-MM-dd}.log"/>
     </rollingPolicy>
 
     <layout class="org.apache.log4j.PatternLayout">
@@ -42,12 +42,12 @@
     </layout>
   </appender>
 
-  <logger name="gobblin.runtime" additivity="false">
+  <logger name="gobblin2.runtime2" additivity="false">
     <level value="INFO"/>
     <appender-ref ref="FileRoll" />
   </logger>
 
-  <logger name="gobblin.runtime.mapreduce.CliMRJobLauncher" additivity="false">
+  <logger name="gobblin2.runtime2.mapreduce.CliMRJobLauncher" additivity="false">
     <level value="ERROR"/>
     <appender-ref ref="Console" />
   </logger>
