+++ gobblin-core/src/main/java/gobblin2/state/ConstructState.java
@@ -1,0 +1,36 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package gobblin2.state;
+
+import java.util.Properties;
+import gobblin2.configuration.State;
+
+/***
+ * Shim layer for org.apache.gobblin.state.ConstructState
+ */
+public class ConstructState extends org.apache.gobblin.state.ConstructState {
+  public ConstructState() {
+  }
+
+  public ConstructState(Properties properties) {
+    super(properties);
+  }
+
+  public ConstructState(State otherState) {
+    super(otherState);
+  }
+}
