+++ gobblin-metrics-libs/gobblin-metrics-base/src/test/resources/performance-testng.xml
@@ -23,7 +23,7 @@
         </run>
     </groups>
     <packages>
-        <package name="org.apache.gobblin.metrics.*" />
+        <package name="org.apache.gobblin2.metrics.*" />
     </packages>
 </test>
 </suite>
