+++ gobblin-oozie/src/test/resources/local/gobblin-oozie-example-workflow.xml
@@ -27,7 +27,7 @@
 					<value>true</value>
 				</property>
 			</configuration>
-			<main-class>gobblin.runtime.local.CliLocalJobLauncher</main-class>
+			<main-class>gobblin2.runtime2.local.CliLocalJobLauncher</main-class>
 			<arg>--jobconfig</arg>
 			<arg>${nameNode}/path/to/jobconfig.properties</arg>
 			<arg>--sysconfig</arg>
