--- gobblin-runtime/src/main/java/gobblin/runtime/TaskState.java
@@ -1,5 +1,0 @@
-package gobblin.runtime;
-
-public class TaskState extends org.apache.gobblin.runtime.TaskState {
-
-}
