+++ gobblin-runtime/src/main/java/gobblin2/runtime2/CheckpointableWatermarkState.java
@@ -1,0 +1,33 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package gobblin2.runtime2;
+
+import com.google.gson.Gson;
+import gobblin2.source.extractor.CheckpointableWatermark;
+
+/***
+ * Shim layer for org.apache.gobblin.runtime.CheckpointableWatermarkState
+ */
+public class CheckpointableWatermarkState extends org.apache.gobblin.runtime.CheckpointableWatermarkState {
+  public CheckpointableWatermarkState(CheckpointableWatermark watermark, Gson gson) {
+    super(watermark, gson);
+  }
+
+  public CheckpointableWatermarkState() {
+    super();
+  }
+}
