+++ gobblin-runtime/src/main/java/gobblin2/runtime2/TaskState.java
@@ -1,0 +1,5 @@
+package gobblin2.runtime2;
+
+public class TaskState extends org.apache.gobblin.runtime.TaskState {
+
+}
