+++ gobblin-runtime/src/test/resources/gobblin2/runtime2/instance/SimpleHelloWorldJob.template
