+++ ligradle/findbugs/findbugsExclude.xml
@@ -46,7 +46,7 @@
   </Match>
   <!-- Ignored until https://github.com/linkedin/gobblin/issues/976 has been resolved -->
   <Match>
-    <Class name="org.apache.gobblin.source.jdbc.JdbcExtractor" />
+    <Class name="org.apache.gobblin2.source.jdbc.JdbcExtractor" />
     <Bug pattern="OBL_UNSATISFIED_OBLIGATION,ODR_OPEN_DATABASE_RESOURCE" />
   </Match>
 </FindBugsFilter>
